+++ by.issoft.paramCollector/src/main/java/org/by/issoft/paramCollector/params/scalarParamValues/FreePhysicalMemoryValue.java
@@ -31,7 +31,7 @@ public class FreePhysicalMemoryValue extends ScalarParamValue<Long> {
 	}
 
 	@Override
-	public Long getDoubleValue() {
+	public Long getLongValue() {
 		Long double1 = new Long(getValue());
 		return double1;
 	}
