+++ by.issoft.paramCollector/src/main/java/org/by/issoft/paramCollector/params/scalarParamValues/PhysicalMemoryUsageValue.java
@@ -28,7 +28,7 @@ public class PhysicalMemoryUsageValue extends ScalarParamValue<Long> {
 	}
 
 	@Override
-	public Long getDoubleValue() {
+	public Long getLongValue() {
 		Long double1 = new Long(getValue());
 		return double1;
 	}
